feat: CPR발생보고서 추가

+++ src/main/java/ai/planit/pev/domain/ods/record/dto/Record.java
@@ -42,7 +42,7 @@ public class Record {
         private String note;
         private int mdfmId;
         private int mdfmFomSeq;
-        private int mdrcId;
+        private long mdrcId;
         private int mdrcFomSeq;
         private String examKey;
         private String pacsImgIptnCd;
